backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
xgboost>=1.7.0
python-dotenv==0.19.0
pymongo==4.1.1
dnspython==2.2.1
spacy==3.6.1
transformers==4.31.0
tensorflow==2.13.0
opencv-python==4.8.0.76
pillow==10.0.0
pytest==7.4.0 